Level for Alpha version, and added intro video
Made a bunch of changes to level and added set dressing. Doors are not set up properly and will need to be looked at. There's something funky with shaders imported from explosions pack, but Aedin says it will work and takes responsibility if it explodes. Also added Alex's intro video.